fix: Split up tests into different files, and move around the includes appropriately.
