+++ lib/NativeCode/DynamicLibraryLoaderHelper/Tests/pch.h
@@ -1,7 +1,12 @@
 //
 // pch.h
 //
-
 #pragma once
+
+#define NOMINMAX
+#ifdef IServiceProvider
+#undef IServiceProvider
+#endif
+#include <msclr/marshal_cppstd.h>
 
 #include "gtest/gtest.h"
